refactor(services): rename base-http.service to http.service
